requirements.txt
ultralytics==8.0.42
uvicorn[standard]==0.20.0
gunicorn==20.1.0
fastapi[all]==0.89.1
Pillow
loguru
pytest
pytest-cov
#yolo==8.1.18

black==24.4.2
flake8==7.0.0